Moved Button and Icon component to journeys/blocks
