+++ apps/journeys/pages/[journeyId].tsx
@@ -71,6 +71,37 @@ export const getServerSideProps: GetServerSideProps<JourneyPageProps> = async (
                 }
               }
             }
+            ... on ButtonBlock {
+              id
+              parentBlockId
+              label
+              variant
+              color
+              size
+              startIcon {
+                name
+                color
+                size
+              }
+              endIcon {
+                name
+                color
+                size
+              }
+              action {
+                __typename
+                gtmEventName
+                ... on NavigateToBlockAction {
+                  blockId
+                }
+                ... on NavigateToJourneyAction {
+                  journeyId
+                }
+                ... on LinkAction {
+                  url
+                }
+              }
+            }
           }
         }
       }
