+++ apps/journeys/src/components/blocks/Icon/Icon.spec.tsx
@@ -4,7 +4,7 @@ import {
   IconColor,
   IconName,
   IconSize
-} from '../../../__generated__/globalTypes'
+} from '../../../../__generated__/globalTypes'
 
 const block: IconProps = {
   __typename: 'Icon',
