--- libs/shared/ui/src/components/Action/actionFields.ts
@@ -1,17 +1,0 @@
-import { gql } from '@apollo/client'
-
-export const ACTION_FIELDS = gql`
-  fragment ActionFields on Action {
-    __typename
-    gtmEventName
-    ... on NavigateAction {
-      gtmEventName
-    }
-    ... on NavigateToJourneyAction {
-      journeyId
-    }
-    ... on LinkAction {
-      url
-    }
-  }
-`
