--- libs/shared/ui/src/components/Action/index.ts
@@ -1,1 +1,0 @@
-export { ACTION_FIELDS } from './actionFields'
